+++ ahcore/utils/types.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
+from enum import Enum
 from typing import Any
 
 import numpy as np
 import numpy.typing as npt
+import torch
 from dlup.data.dataset import Dataset
 from pydantic import AfterValidator
 from typing_extensions import Annotated
@@ -29,3 +31,35 @@ GenericArray = npt.NDArray[np.generic]
 
 DlupDatasetSample = dict[str, Any]
 _DlupDataset = Dataset[DlupDatasetSample]
+
+
+class NormalizationType(str, Enum):
+    SIGMOID = "sigmoid"
+    SOFTMAX = "softmax"
+    LOGITS = "logits"
+
+    def normalize(self):
+        if self == NormalizationType.SIGMOID:
+            return torch.sigmoid
+        elif self == NormalizationType.SOFTMAX:
+            return torch.softmax
+        elif self == NormalizationType.LOGITS:
+            return lambda x: x
+        else:
+            raise ValueError("Function not supported")
+
+
+class InferencePrecision(str, Enum):
+    FP16 = "float16"
+    FP32 = "float32"
+    UINT8 = "uint8"
+
+    def get_multiplier(self) -> float:
+        if self == InferencePrecision.FP16:
+            return 1.0
+        elif self == InferencePrecision.FP32:
+            return 1.0
+        elif self == InferencePrecision.UINT8:
+            return 255.0
+        else:
+            raise NotImplementedError(f"Precision {self} is not supported for {self.__class__.__name__}.")
